change test double builders
